Move channel side effect tests to module file
